pyproject.toml
[tool.poetry]
name = "spotify"
version = "0.1.0"
description = ""
authors = ["Emre Can <emredjan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.23.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
black = "^21.9b0"
ipython = "^7.28.0"
pandas = "^1.3.3"
spotipy = "^2.19.0"
click = "^8.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
